strict: import the script surface from "vitexec" instead of a separate entry
The import only carries types and is stripped before the script runs, so the
main entry re-exports it and the ./strict export goes away.
