+++ packages/vitexec/package.json
@@ -21,10 +21,6 @@
     "./cli": {
       "types": "./dist/cli.d.ts",
       "import": "./dist/cli.js"
-    },
-    "./strict": {
-      "types": "./dist/strict/api.d.ts",
-      "import": "./dist/strict/api.js"
     }
   },
   "bin": {
